LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/history_go_undefined-expected.txt

main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)
main frame - has 1 onunload handler(s)

FAIL history.forward() with session history assert_array_equals: Pages opened during history navigation property 2, expected 2 but got 3
